LayoutTests/imported/w3c/web-platform-tests/css/css-animations/event-order.tentative-expected.txt


FAIL Same events are ordered by elements assert_equals: Event #1 pseudoElements should match expected (string) "" but got (undefined) undefined
FAIL Same events on pseudo-elements follow the prescribed order assert_equals: Event #1 pseudoElements should match expected (string) "" but got (undefined) undefined
FAIL Start and iteration events are ordered by time assert_equals: Event #1 types should match (expected: animationiteration, animationstart, actual: animationstart, animationiteration) expected "animationiteration" but got "animationstart"
FAIL Iteration and end events are ordered by time assert_equals: Event #1 types should match (expected: animationiteration, animationend, actual: animationend, animationiteration) expected "animationiteration" but got "animationend"
FAIL Start and end events are sorted correctly when fired simultaneously assert_equals: Event #1 targets should match expected Element node <div style="animation: anim 100s 2"></div> but got Element node <div style="animation: anim 100s 100s"></div>
